fix copilot review
