+++ src/app/shared/import/domain/import-contracts.interfaces.ts
@@ -38,28 +38,38 @@ export type ImportPipelineStage =
 /**
  * Canonical error codes emitted by the generic import pipeline.
  *
- * Adapters may emit additional context-specific codes
- * (e.g. `'CABLE_NOT_FOUND'` for the Study adapter), hence `ImportErrorCode`
- * is typed as a plain `string`.
+ * The runtime constant `IMPORT_ERROR_CODES` lives in `import-contracts.constantes.ts`;
+ * this type is derived from it for compile-time checking and autocomplete.
+ * Adapters may emit additional context-specific codes via
+ * {@link AdapterImportErrorCode} (e.g. `'CABLE_NOT_FOUND'`).
  */
-export const IMPORT_ERROR_CODES = {
-  FILE_TYPE_NOT_ALLOWED: 'FILE_TYPE_NOT_ALLOWED',
-  FILE_READ_ERROR: 'FILE_READ_ERROR',
-  FILE_DECODE_ERROR: 'FILE_DECODE_ERROR',
-  FILE_PARSE_ERROR: 'FILE_PARSE_ERROR',
-  VALIDATION_ERROR: 'VALIDATION_ERROR',
-  MAPPING_ERROR: 'MAPPING_ERROR',
-  PERSISTENCE_ERROR: 'PERSISTENCE_ERROR',
-  UUID_COLLISION_REJECTED: 'UUID_COLLISION_REJECTED'
-} as const;
+export type CanonicalImportErrorCode =
+  | 'FILE_TYPE_NOT_ALLOWED'
+  | 'FILE_READ_ERROR'
+  | 'FILE_DECODE_ERROR'
+  | 'FILE_PARSE_ERROR'
+  | 'VALIDATION_ERROR'
+  | 'MAPPING_ERROR'
+  | 'PERSISTENCE_ERROR'
+  | 'UUID_COLLISION_REJECTED';
+
+/**
+ * Branded string carrying an adapter-specific error code.
+ *
+ * Adapters opt into the open extensibility surface by tagging their code
+ * via the `adapterErrorCode` helper (see `import-contracts.constantes.ts`),
+ * which keeps autocomplete intact for canonical codes while still allowing
+ * arbitrary string literals from adapters.
+ */
+export type AdapterImportErrorCode = string & { readonly __adapterImportErrorCode: unique symbol };
 
 /**
  * Standardized error code for the import pipeline.
  *
- * See {@link IMPORT_ERROR_CODES} for the canonical set of values.
- * Adapters may emit context-specific codes as additional string literals.
+ * Either a {@link CanonicalImportErrorCode} (autocompleted, type-checked) or an
+ * {@link AdapterImportErrorCode} (branded string for adapter-specific codes).
  */
-export type ImportErrorCode = string;
+export type ImportErrorCode = CanonicalImportErrorCode | AdapterImportErrorCode;
 
 /**
  * A structured import error carrying the pipeline stage it originated from.
